requirements.txt
pillow==10.3.0
pypng==0.20220715.0
stepic==0.5.0
